requirements.txt
flask==3.1.0
flask_socketio==5.4.1
flask-cors==4.0.0
pydantic==2.10.1
PyYAML==6.0.1
requests==2.31.0
numpy==1.26.4
pyautogui==0.9.53
pycryptodome==3.20.0